+++ daemon/started_service.go
@@ -62,6 +62,8 @@ type StartedService struct {
 	runtimeSequence          atomic.Uint64
 	clashModeSubscriber      *observable.Subscriber[struct{}]
 	clashModeObserver        *observable.Observer[struct{}]
+	trafficSubscriber        *observable.Subscriber[struct{}]
+	trafficObserver          *observable.Observer[struct{}]
 
 	connectionEventSubscriber *observable.Subscriber[trafficontrol.ConnectionEvent]
 	connectionEventObserver   *observable.Observer[trafficontrol.ConnectionEvent]
@@ -107,6 +109,7 @@ func NewStartedService(options ServiceOptions) *StartedService {
 		urlTestSessionByGroup:     make(map[string]string),
 		urlTestSessionSubscriber:  observable.NewSubscriber[struct{}](1),
 		clashModeSubscriber:       observable.NewSubscriber[struct{}](1),
+		trafficSubscriber:         observable.NewSubscriber[struct{}](1),
 		connectionEventSubscriber: observable.NewSubscriber[trafficontrol.ConnectionEvent](256),
 	}
 	s.serviceStatusObserver = observable.NewObserver(s.serviceStatusSubscriber, 2)
@@ -114,6 +117,7 @@ func NewStartedService(options ServiceOptions) *StartedService {
 	s.urlTestObserver = observable.NewObserver(s.urlTestSubscriber, 1)
 	s.urlTestSessionObserver = observable.NewObserver(s.urlTestSessionSubscriber, 1)
 	s.clashModeObserver = observable.NewObserver(s.clashModeSubscriber, 1)
+	s.trafficObserver = observable.NewObserver(s.trafficSubscriber, 1)
 	s.connectionEventObserver = observable.NewObserver(s.connectionEventSubscriber, 64)
 	return s
 }
@@ -209,7 +213,9 @@ func (s *StartedService) StartOrReloadService(profileContent string, options *Ov
 	instance.urlTestHistoryStorage.SetHook(s.urlTestSubscriber)
 	if instance.clashServer != nil {
 		instance.clashServer.SetModeUpdateHook(s.clashModeSubscriber)
-		instance.clashServer.(*clashapi.Server).TrafficManager().SetEventHook(s.connectionEventSubscriber)
+		trafficManager := instance.clashServer.(*clashapi.Server).TrafficManager()
+		trafficManager.SetEventHook(s.connectionEventSubscriber)
+		trafficManager.SetTrafficHook(s.trafficSubscriber)
 	}
 	s.serviceAccess.Unlock()
 	err = instance.Start()
@@ -238,6 +244,7 @@ func (s *StartedService) Close() {
 	s.urlTestSubscriber.Close()
 	s.urlTestSessionSubscriber.Close()
 	s.clashModeSubscriber.Close()
+	s.trafficSubscriber.Close()
 	s.connectionEventSubscriber.Close()
 }
 
@@ -677,6 +684,11 @@ func (s *StartedService) SubscribeConnections(request *SubscribeConnectionsReque
 		return err
 	}
 	defer s.connectionEventObserver.UnSubscribe(subscription)
+	trafficSubscription, trafficDone, err := s.trafficObserver.Subscribe()
+	if err != nil {
+		return err
+	}
+	defer s.trafficObserver.UnSubscribe(trafficSubscription)
 
 	connectionSnapshots := make(map[uuid.UUID]connectionSnapshot)
 	initialEvents := s.buildInitialConnectionState(trafficManager, connectionSnapshots)
@@ -692,8 +704,13 @@ func (s *StartedService) SubscribeConnections(request *SubscribeConnectionsReque
 	if interval <= 0 {
 		interval = time.Second
 	}
-	ticker := time.NewTicker(interval)
-	defer ticker.Stop()
+	var trafficTimer *time.Timer
+	var trafficTimerC <-chan time.Time
+	defer func() {
+		if trafficTimer != nil {
+			trafficTimer.Stop()
+		}
+	}()
 
 	for {
 		select {
@@ -702,6 +719,8 @@ func (s *StartedService) SubscribeConnections(request *SubscribeConnectionsReque
 		case <-server.Context().Done():
 			return server.Context().Err()
 		case <-done:
+			return nil
+		case <-trafficDone:
 			return nil
 
 		case event := <-subscription:
@@ -727,7 +746,16 @@ func (s *StartedService) SubscribeConnections(request *SubscribeConnectionsReque
 				}
 			}
 
-		case <-ticker.C:
+		case <-trafficSubscription:
+			if trafficTimerC != nil {
+				continue
+			}
+			trafficTimer = time.NewTimer(interval)
+			trafficTimerC = trafficTimer.C
+			continue
+
+		case <-trafficTimerC:
+			trafficTimerC = nil
 			protoEvents := s.buildTrafficUpdates(trafficManager, connectionSnapshots)
 			if len(protoEvents) == 0 {
 				continue